fix: remove supported from autogen status (#3714)
