+++ config/crds/kyverno.io_policies.yaml
@@ -2467,11 +2467,6 @@ spec:
                     items:
                       type: string
                     type: array
-                  supported:
-                    description: Supported indicates the autogen supported controllers
-                    items:
-                      type: string
-                    type: array
                 type: object
               conditions:
                 description: Conditions is a list of conditions that apply to the
