+++ config/install_debug.yaml
@@ -2471,11 +2471,6 @@ spec:
                     items:
                       type: string
                     type: array
-                  supported:
-                    description: Supported indicates the autogen supported controllers
-                    items:
-                      type: string
-                    type: array
                 type: object
               conditions:
                 description: Conditions is a list of conditions that apply to the
@@ -8289,11 +8284,6 @@ spec:
                     type: array
                   requested:
                     description: Requested indicates the autogen requested controllers
-                    items:
-                      type: string
-                    type: array
-                  supported:
-                    description: Supported indicates the autogen supported controllers
                     items:
                       type: string
                     type: array
