+++ docs/crd/v1/index.html
@@ -730,7 +730,7 @@ attributes for keyless verification, or a nested attestor declaration.</p>
 </p>
 <p>
 <p>AutogenStatus contains autogen status information.
-It indicates requested, supported and effective autogen controllers used when
+It indicates requested and effective autogen controllers used when
 automatically generating rules.</p>
 </p>
 <table class="table table-striped">
@@ -750,17 +750,6 @@ automatically generating rules.</p>
 </td>
 <td>
 <p>Requested indicates the autogen requested controllers</p>
-</td>
-</tr>
-<tr>
-<td>
-<code>supported</code></br>
-<em>
-[]string
-</em>
-</td>
-<td>
-<p>Supported indicates the autogen supported controllers</p>
 </td>
 </tr>
 <tr>
